chore: extract /contract-info route handler
